LayoutTests/accessibility/ios-simulator/text-role.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="textrole" tabindex=0 role="text" aria-label="all at once">
<b>hello world</b>
<i>this is a test</i>
<a href="#">more test</a>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test ensures that the \"text\" role works as expected. That is, it creates a static text object that can be controlled with aria-label.");

    if (window.accessibilityController) {

        var textrole = accessibilityController.accessibleElementById("textrole");
        shouldBe("textrole.description", "'AXLabel: all at once'");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
